+++ shuup/admin/modules/orders/views/__init__.py
@@ -15,7 +15,10 @@ from .payment import (
     OrderCreatePaymentView, OrderDeletePaymentView, OrderSetPaidView
 )
 from .refund import OrderCreateFullRefundView, OrderCreateRefundView
-from .shipment import OrderCreateShipmentView, ShipmentDeleteView
+from .shipment import (
+    OrderCreateShipmentView, ShipmentDeleteView, ShipmentListView,
+    ShipmentSetSentView
+)
 from .status import OrderStatusEditView, OrderStatusListView
 
 __all__ = [
@@ -23,5 +26,5 @@ __all__ = [
     "OrderListView", "OrderCreatePaymentView", "OrderCreateFullRefundView",
     "OrderCreateRefundView", "OrderCreateShipmentView", "OrderSetPaidView",
     "OrderSetStatusView", "OrderStatusEditView", "OrderStatusListView", "ShipmentDeleteView",
-    "UpdateAdminCommentView", "OrderDeletePaymentView"
+    "UpdateAdminCommentView", "OrderDeletePaymentView", "ShipmentSetSentView", "ShipmentListView"
 ]