+++ shuup/admin/modules/orders/views/list.py
@@ -25,10 +25,10 @@ from shuup.utils.i18n import format_money, get_locally_formatted_datetime
 class OrderListView(PicotableListView):
     model = Order
     default_columns = [
-        Column("identifier", _(u"Order"), linked=True, filter_config=TextFilter(operator="startswith")),
-        Column("order_date", _(u"Order Date"), display="format_order_date", filter_config=DateRangeFilter()),
+        Column("identifier", _("Order"), linked=True, filter_config=TextFilter(operator="startswith")),
+        Column("order_date", _("Order Date"), display="format_order_date", filter_config=DateRangeFilter()),
         Column(
-            "customer", _(u"Customer"), display="format_customer_name",
+            "customer", _("Customer"), display="format_customer_name",
             filter_config=MultiFieldTextFilter(
                 filter_fields=(
                     "customer__email", "customer__name", "billing_address__name",
@@ -37,12 +37,12 @@ class OrderListView(PicotableListView):
             )
         ),
         Column(
-            "status", _(u"Status"), filter_config=ChoicesFilter(choices=OrderStatus.objects.all()),
+            "status", _("Status"), filter_config=ChoicesFilter(choices=OrderStatus.objects.all()),
         ),
-        Column("payment_status", _(u"Payment Status"), filter_config=ChoicesFilter(choices=PaymentStatus.choices)),
-        Column("shipping_status", _(u"Shipping Status"), filter_config=ChoicesFilter(choices=ShippingStatus.choices)),
+        Column("payment_status", _("Payment Status"), filter_config=ChoicesFilter(choices=PaymentStatus.choices)),
+        Column("shipping_status", _("Shipping Status"), filter_config=ChoicesFilter(choices=ShippingStatus.choices)),
         Column(
-            "taxful_total_price_value", _(u"Total"), sort_field="taxful_total_price_value",
+            "taxful_total_price_value", _("Total"), sort_field="taxful_total_price_value",
             display="format_taxful_total_price", class_name="text-right",
             filter_config=RangeFilter(field_type="number", filter_field="taxful_total_price_value")
         ),
@@ -88,6 +88,6 @@ class OrderListView(PicotableListView):
     def get_object_abstract(self, instance, item):
         return [
             {"text": "%s" % instance, "class": "header"},
-            {"title": _(u"Total"), "text": item.get("taxful_total_price_value")},
-            {"title": _(u"Status"), "text": item.get("status")}
+            {"title": _("Total"), "text": item.get("taxful_total_price_value")},
+            {"title": _("Status"), "text": item.get("status")}
         ]