+++ shuup/admin/modules/orders/views/shipment.py
@@ -11,17 +11,25 @@ import six
 from django import forms
 from django.contrib import messages
 from django.http.response import HttpResponseRedirect
+from django.template.loader import render_to_string
 from django.utils.translation import ugettext as _
 from django.views.generic import DetailView, UpdateView
 
 from shuup.admin.form_modifier import ModifiableFormMixin, ModifiableViewMixin
+from shuup.admin.supplier_provider import get_supplier
 from shuup.admin.toolbar import PostActionButton, Toolbar
 from shuup.admin.utils.forms import add_form_errors_as_messages
-from shuup.admin.utils.urls import get_model_url
+from shuup.admin.utils.picotable import Column
+from shuup.admin.utils.urls import get_model_url, NoModelUrl
+from shuup.admin.utils.views import PicotableListView
 from shuup.core.excs import (
     NoProductsToShipException, NoShippingAddressException
 )
-from shuup.core.models import Order, Product, Shipment, Shop
+from shuup.core.models import (
+    Order, Product, Shipment, ShipmentProduct, ShipmentStatus, Shop,
+    UnitInterface
+)
+from shuup.utils.django_compat import reverse
 from shuup.utils.excs import Problem
 
 
@@ -30,6 +38,7 @@ class ShipmentForm(ModifiableFormMixin, forms.Form):
 
     description = forms.CharField(required=False)
     tracking_code = forms.CharField(required=False)
+    tracking_url = forms.URLField(required=False, label=_("Tracking URL"))
 
 
 class OrderCreateShipmentView(ModifiableViewMixin, UpdateView):
@@ -146,6 +155,7 @@ class OrderCreateShipmentView(ModifiableViewMixin, UpdateView):
             order=order,
             supplier_id=self._get_supplier_id(),
             tracking_code=form.cleaned_data.get("tracking_code"),
+            tracking_url=form.cleaned_data.get("tracking_url"),
             description=form.cleaned_data.get("description"))
         has_extension_errors = self.form_valid_hook(form, unsaved_shipment)
 
@@ -191,3 +201,102 @@ class ShipmentDeleteView(DetailView):
         shipment.soft_delete()
         messages.success(request, _("Shipment %s has been deleted.") % shipment.pk)
         return HttpResponseRedirect(self.get_success_url())
+
+
+class ShipmentSetSentView(DetailView):
+    model = Shipment
+    context_object_name = "shipment"
+
+    def get_queryset(self):
+        shop_ids = Shop.objects.get_for_user(self.request.user).values_list("id", flat=True)
+        return Shipment.objects.filter(order__shop_id__in=shop_ids)
+
+    def get_success_url(self):
+        if self.request.GET.get("next"):
+            return self.request.GET["next"]
+        return get_model_url(self.get_object().order)
+
+    def get(self, request, *args, **kwargs):
+        return HttpResponseRedirect(self.get_success_url())
+
+    def post(self, request, *args, **kwargs):
+        shipment = self.get_object()
+        shipment.set_sent()
+        messages.success(request, _("Shipment has been marked as sent."))
+        return HttpResponseRedirect(self.get_success_url())
+
+
+class ShipmentListView(PicotableListView):
+    model = Shipment
+    template_name = "shuup/admin/orders/shipment_picotable.jinja"
+    default_columns = [
+        Column("id", _("ID"), linked=True),
+        Column("order", _("Order"), display="get_order", raw=True),
+        Column("get_content", _("Content"), display="get_content", raw=True),
+        Column("supplier", _("Supplier"), display="get_supplier", raw=True),
+        Column("tracking_code", _("Tracking Code"), display="tracking_code_url", raw=True),
+        Column("status", _("Status"), display="create_action_buttons", raw=True),
+    ]
+
+    def tracking_code_url(self, instance):
+        if instance.tracking_url:
+            return f'<a target="_blank" ref="noopened" href="{instance.tracking_url}">{instance.tracking_code}</a>'
+        return instance.tracking_code
+
+    def get_supplier(self, instance):
+        return instance.supplier.name
+
+    def get_order(self, instance):
+        try:
+            order_url = get_model_url(instance.order)
+            return f'<a href="{order_url}">{str(instance.order)}</a>'
+        except NoModelUrl:
+            return str(instance.order)
+
+    def get_content(self, instance):
+        shipment_products = ShipmentProduct.objects.filter(shipment=instance).select_related(
+            "product", "product__sales_unit"
+        )
+        content = []
+
+        for shipment_product in shipment_products:
+            unit = UnitInterface(shipment_product.product.sales_unit)
+            quantity = unit.render_quantity(shipment_product.quantity)
+            content.append(f"{shipment_product.product} ({quantity})")
+        return ", ".join(content)
+
+    def create_action_buttons(self, instance):
+        if instance.order.shipping_method and instance.order.shipping_method.carrier.uses_default_shipments_manager:
+            if instance.status not in (ShipmentStatus.SENT, ShipmentStatus.ERROR):
+                url = "{base_url}?next={next_url}".format(
+                    base_url=reverse("shuup_admin:order.set-shipment-sent", kwargs={"pk": instance.pk}),
+                    next_url=reverse('shuup_admin:order.shipments.list')
+                )
+                return render_to_string("shuup/admin/orders/_set_shipments_status_button.jinja", {
+                    "shipment_id": instance.pk,
+                    "url": url,
+                })
+        return instance.status.label
+
+    def __init__(self):
+        super().__init__()
+        self.columns = self.default_columns
+
+    def get_queryset(self):
+        queryset = super().get_queryset().exclude(status=ShipmentStatus.DELETED).select_related(
+            "order", "order__shipping_method"
+        )
+        supplier = get_supplier(self.request)
+        if supplier:
+            queryset = queryset.filter(supplier=supplier)
+        return queryset
+
+    def get_object_abstract(self, instance, item):
+        return [
+            {"text": _("Shipment {shipment_id}").format(shipment_id=item.get("id")), "class": "header"},
+            {"title": _("Order"), "text": " ", "raw": item.get("order")},
+            {"title": _("Supplier"), "text": item.get("supplier")},
+            {"title": _("Content"), "text": item.get("get_content")},
+            {"title": _("Tracking code"), "text": " ", "raw": item.get("tracking_code")},
+            {"title": _("Status"), "text": " ", "raw": item.get("status")}
+        ]