+++ shuup/core/models/_shipments.py
@@ -19,7 +19,7 @@ from shuup.core.fields import (
     InternalIdentifierField, MeasurementField, QuantityField
 )
 from shuup.core.models import ShuupModel
-from shuup.core.signals import shipment_deleted
+from shuup.core.signals import shipment_deleted, shipment_sent
 from shuup.core.utils.units import get_shuup_volume_unit
 from shuup.utils.analog import define_log_model
 
@@ -34,11 +34,11 @@ class ShipmentStatus(Enum):
     DELETED = 20
 
     class Labels:
-        NOT_SENT = _("not sent")
-        SENT = _("sent")
-        RECEIVED = _("received")
-        ERROR = _("error")
-        DELETED = _("deleted")
+        NOT_SENT = _("Not sent")
+        SENT = _("Sent")
+        RECEIVED = _("Received")
+        ERROR = _("Error")
+        DELETED = _("Deleted")
 
 
 class ShipmentType(Enum):
@@ -50,10 +50,12 @@ class ShipmentType(Enum):
         IN = _("incoming")
 
 
-class ShipmentManager(models.Manager):
-
+class ShipmentQueryset(models.QuerySet):
     def all_except_deleted(self, language=None, shop=None):
         return self.exclude(status=ShipmentStatus.DELETED)
+
+    def sent(self):
+        return self.filter(status=ShipmentStatus.SENT)
 
 
 class Shipment(ShuupModel):
@@ -66,6 +68,7 @@ class Shipment(ShuupModel):
     created_on = models.DateTimeField(auto_now_add=True, verbose_name=_("created on"))
     status = EnumIntegerField(ShipmentStatus, default=ShipmentStatus.NOT_SENT, verbose_name=_("status"))
     tracking_code = models.CharField(max_length=64, blank=True, verbose_name=_("tracking code"))
+    tracking_url = models.URLField(blank=True, verbose_name=_("tracking url"))
     description = models.CharField(max_length=255, blank=True, verbose_name=_("description"))
     volume = MeasurementField(
         unit=get_shuup_volume_unit(),
@@ -77,9 +80,8 @@ class Shipment(ShuupModel):
     )
     identifier = InternalIdentifierField(unique=True)
     type = EnumIntegerField(ShipmentType, default=ShipmentType.OUT, verbose_name=_("type"))
-    # TODO: documents = models.ManyToManyField(FilerFile)
 
-    objects = ShipmentManager()
+    objects = ShipmentQueryset.as_manager()
 
     class Meta:
         verbose_name = _('shipment')
@@ -122,6 +124,9 @@ class Shipment(ShuupModel):
     def is_deleted(self):
         return bool(self.status == ShipmentStatus.DELETED)
 
+    def is_sent(self):
+        return bool(self.status == ShipmentStatus.SENT)
+
     def cache_values(self):
         """
         (Re)cache `.volume` and `.weight` for this Shipment from within the ShipmentProducts.
@@ -137,6 +142,17 @@ class Shipment(ShuupModel):
     @property
     def total_products(self):
         return (self.products.aggregate(quantity=models.Sum("quantity"))["quantity"] or 0)
+
+    def set_sent(self):
+        """
+        Mark the shipment as sent.
+        """
+        if self.status == ShipmentStatus.SENT:
+            return
+
+        self.status = ShipmentStatus.SENT
+        self.save()
+        shipment_sent.send(sender=type(self), order=self.order, shipment=self)
 
     def set_received(self, purchase_prices=None, created_by=None):
         """