+++ shuup/gdpr/models.py
@@ -10,7 +10,6 @@ from __future__ import unicode_literals
 from django.conf import settings
 from django.db import models
 from django.utils.encoding import python_2_unicode_compatible
-from django.utils.lru_cache import lru_cache
 from django.utils.translation import activate, get_language
 from django.utils.translation import ugettext_lazy as _
 from parler.models import TranslatableModel, TranslatedFields
@@ -18,11 +17,12 @@ from reversion.models import Version
 
 from shuup.gdpr.utils import get_active_consent_pages
 from shuup.simple_cms.models import Page
+from shuup.utils.i18n import lang_lru_cache
 
 GDPR_ANONYMIZE_TASK_TYPE_IDENTIFIER = "gdpr_anonymize"
 
 
-@lru_cache()
+@lang_lru_cache
 def get_setting(shop):
     instance, created = GDPRSettings.objects.get_or_create(shop=shop)
     if created or not instance.safe_translation_getter("cookie_banner_content"):